blog_api/requirements.txt
django>=4.2
djangorestframework>=3.14
asgiref>=3.7
